+++ built/jtree.node.js
@@ -6,6 +6,10 @@ const GrammarProgram_1 = require("./grammar/GrammarProgram");
 class jtreeNode extends jtree_1.default {
 }
 jtreeNode.executeFile = (programPath, grammarPath) => jtreeNode.makeProgram(programPath, grammarPath).execute(programPath);
+jtreeNode.executeFiles = (programPaths, grammarPath) => {
+    const programConstructor = jtreeNode.getProgramConstructor(grammarPath);
+    return programPaths.map(programPath => new programConstructor(fs.readFileSync(programPath, "utf8")).execute(programPath));
+};
 jtreeNode.executeFileSync = (programPath, grammarPath) => jtreeNode.makeProgram(programPath, grammarPath).executeSync(programPath);
 // returns AbstractRuntimeProgram
 jtreeNode.makeProgram = (programPath, grammarPath) => {